Update Accuracy.py
Corrected ROI from 180 radius to 180 diameter :|

Updated to find max bias at or below NECRp.

Repeats the calculations for the central 80% FOV. Purely done to meet the ANZSNM TSC requirements https://www.anzsnm.org.au/documents/item/97

Note: The lower values (below 10% NECRp) are ignored to remove noisy points.

Slight modification to Response plot to match the Y axis of the Siemens results. Note the image voxels are in Bq/ml and the response plot sums these voxels.  The Siemens plot is labelled in cps which appears to be incorrect.